+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Roles and permissions concept ([#12], [#22])
 - Added page to create and view room, start/join meetings, change settings ([#4], [#18],[#32],[#33])
 - Added file upload for room owner and file download for participants ([#4], [#18])
-- Added setting menu for administrators ([#35])
+- Added setting menu for administrators ([#35], [#38])
 
 [#1]: https://github.com/THM-Health/PILOS/issues/1
 [#3]: https://github.com/THM-Health/PILOS/pull/3
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#22]: https://github.com/THM-Health/PILOS/pull/22
 [#32]: https://github.com/THM-Health/PILOS/issues/32
 [#33]: https://github.com/THM-Health/PILOS/pull/33
-[#33]: https://github.com/THM-Health/PILOS/pull/35
+[#35]: https://github.com/THM-Health/PILOS/pull/35
+[#38]: https://github.com/THM-Health/PILOS/issues/38
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/3c8359cdb0395546fe97aeabf1a40f93002b182c...HEAD